# server/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: JobBoard API
  description: Full-featured Job Board Platform API
  version: 1.0.0
  contact:
    name: JobBoard Team
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: http://localhost/api
    description: Docker production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [ADMIN, EMPLOYER, SEEKER]
        avatarUrl:
          type: string
          nullable: true
        resumeUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName, role]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                firstName:
                  type: string
                lastName:
                  type: string
                role:
                  type: string
                  enum: [EMPLOYER, SEEKER]
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    $ref: '#/components/schemas/AuthTokens'
        '400':
          description: Validation error
        '409':
          description: Email already exists
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  tokens:
                    $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
        '401':
          description: Invalid refresh token
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
        '401':
          description: Not authenticated
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logged out successfully
  /jobs:
    get:
      tags: [Jobs]
      summary: List approved jobs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of approved jobs
    post:
      tags: [Jobs]
      summary: Create a new job
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, description, location, type, companyId]
              properties:
                title:
                  type: string
                description:
                  type: string
                location:
                  type: string
                type:
                  type: string
                  enum: [FULL_TIME, PART_TIME, CONTRACT, INTERNSHIP, REMOTE]
                salaryMin:
                  type: integer
                salaryMax:
                  type: integer
                requirements:
                  type: string
                companyId:
                  type: string
      responses:
        '201':
          description: Job created (pending approval)
        '401':
          description: Not authenticated
        '403':
          description: Not authorized (seekers cannot post jobs)
  /jobs/search:
    get:
      tags: [Jobs]
      summary: Search and filter jobs
      parameters:
        - name: keyword
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [FULL_TIME, PART_TIME, CONTRACT, INTERNSHIP, REMOTE]
        - name: salaryMin
          in: query
          schema:
            type: integer
        - name: salaryMax
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, salary]
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Search results
  /jobs/{id}:
    get:
      tags: [Jobs]
      summary: Get job details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job details
        '404':
          description: Job not found
    put:
      tags: [Jobs]
      summary: Update a job
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                location:
                  type: string
                type:
                  type: string
                salaryMin:
                  type: integer
                salaryMax:
                  type: integer
                requirements:
                  type: string
      responses:
        '200':
          description: Job updated
        '403':
          description: Not authorized
        '404':
          description: Job not found
    delete:
      tags: [Jobs]
      summary: Delete a job
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job deleted
        '403':
          description: Not authorized
        '404':
          description: Job not found
...
  /applications/job/{jobId}:
    get:
      tags: [Applications]
      summary: List applicants for a job (employer)
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of applicants
        '403':
          description: Not authorized
...